requirements.txt
langchain
langchain-openai
langchain-deepseek
pydantic-settings
loguru
python-dotenv
streamlit
PyYAML
